Implements chunks mode to support large data
