+++ src/main/java/com/upgrade/tools/convert/constants/SchemeConverterSupportedTypes.java
@@ -1,0 +1,12 @@
+package com.upgrade.tools.convert.constants;
+
+/**
+ * @author Albert Gomes Cabral
+ */
+public class SchemeConverterSupportedTypes {
+
+    public static final String POSTGRES = "postgresql";
+
+    public static final String MYSQL = "mysql";
+
+}
