+++ src/main/java/com/upgrade/tools/util/Print.java
@@ -9,16 +9,16 @@ import java.util.regex.Pattern;
  */
 public class Print {
 
-    public static void info(String word) {
-        _print(_GREEN, word);
+    public static void info(String describe) {
+        _print(_COLOR_GREEN, describe);
     }
 
-    public static void error(String word, String ...cause) {
-        if (cause == null) {
-            _print(_RED, word);
+    public static void error(String describe, String ...cause) {
+        if (cause == null || cause.length == 0) {
+            _print(_COLOR_RED, describe);
         }
         else {
-            _print(_RED, word + _BREAK_LINE + Arrays.toString(cause));
+            _print(_COLOR_RED, describe + _BREAK_LINE + Arrays.toString(cause));
         }
     }
 
@@ -28,25 +28,24 @@ public class Print {
         System.out.println("Applying pattern " + pattern.pattern());
 
         System.out.println(
-                "Replace\n" + _LIGHT_BLUE + oldContent + _RESET +
-                        _BREAK_LINE + "By\n" + _GREEN + newContent +
-                            _RESET + _DOUBLE_BREAK_LINE);
-
+            "Replace\n" + _COLOR_LIGHT_BLUE + oldContent + _RESET +
+                    _BREAK_LINE + "By\n" + _COLOR_GREEN + newContent +
+                        _RESET + _DOUBLE_BREAK_LINE);
     }
 
-    public static void warn(String word, String cause) {
+    public static void warn(String describe, Exception cause) {
         if (cause == null) {
-            _print(_YELLOW, word);
+            _print(_COLOR_YELLOW, describe);
         }
         else {
-            _print(_YELLOW, word + _BREAK_LINE +  cause);
+            _print(_COLOR_YELLOW, describe + _BREAK_LINE +  cause);
         }
     }
 
-    private static void _print(String color, String word) {
+    private static void _print(String color, String describe) {
         System.out.println(
-                Instant.now() + _STRING_WHITE_SPACE + color
-                        + _STRING_WHITE_SPACE + word + _RESET);
+            Instant.now() + _STRING_WHITE_SPACE + color
+                + _STRING_WHITE_SPACE + describe + _RESET);
     }
 
     // utilities variables
@@ -59,14 +58,14 @@ public class Print {
 
     // colors variables
 
-    private static final String _GREEN = "\u001B[32m";
+    private static final String _COLOR_GREEN = "\u001B[32m";
 
-    private static final String _LIGHT_BLUE = "\u001B[94m";
+    private static final String _COLOR_LIGHT_BLUE = "\u001B[94m";
 
-    private static final String _RED = "\u001B[31m";
+    private static final String _COLOR_RED = "\u001B[31m";
+
+    private static final String _COLOR_YELLOW = "\u001B[33m";
 
     private static final String _RESET = "\u001B[0m";
-
-    private static final String _YELLOW = "\u001B[33m";
 
 }
