+++ src/main/resources/properties/scheme-converter.properties
@@ -1,0 +1,4 @@
+database.type=postgresql
+new.file.name=sapphire-postgres-20250409.sql
+source.file.name=
+target.file.name=
